kameo-embed-generator-main/kameo-embed-generator-main/requirements.txt
requests
boto3
botocore
python-dotenv